% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterKmeans.se.R
\name{clusterKmeans.se}
\alias{clusterKmeans.se}
\title{k-means clustering of cells}
\usage{
clusterKmeans.se(
  x,
  k,
  more.kmeans.args = list(),
  reddim.type = "PCA",
  output.name = "clusters",
  meta.name = NULL
)
}
\arguments{
\item{x}{A \link[SingleCellExperiment]{SingleCellExperiment} object or one of its subclasses.
Rows correspond to genomic features and columns correspond to cells.}

\item{k}{Arguments to be passed to \code{\link[scrapper]{runKmeans}}.}

\item{more.kmeans.args}{Named list of further arguments to be passed to \code{\link[scrapper]{runKmeans}}.}

\item{reddim.type}{Integer or string specifying the existing embedding in the \code{\link[SingleCellExperiment]{reducedDim}} of \code{x}.}

\item{output.name}{String containing the name of the \code{\link[SummarizedExperiment]{colData}} column in which to store the cluster assignments.}

\item{meta.name}{String containing the name of the \code{\link[SummarizedExperiment]{metadata}} entry in which to store extra clustering output.
If \code{NULL}, no extra clustering output is stored.}
}
\value{
\code{x} is returned with the cluster assignment for each cell stored in the \code{colData}.
Additional clustering output is stored in the \code{metadata}.
}
\description{
Perform k-means clustering on an existing low-dimensional embedding.
}
\examples{
sce <- getTestRnaData.se("pca")
sce <- clusterKmeans.se(sce, k=10)
table(sce$clusters)

}
\seealso{
\code{\link[scrapper]{clusterKmeans}} from the \pkg{scrapper} package.
}
\author{
Aaron Lun
}
