% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctMnn.se.R
\name{correctMnn.se}
\alias{correctMnn.se}
\title{MNN correction}
\usage{
correctMnn.se(
  x,
  block,
  more.mnn.args = list(),
  reddim.type = "PCA",
  output.name = "MNN",
  delayed.transpose = FALSE
)
}
\arguments{
\item{x}{A \link[SingleCellExperiment]{SingleCellExperiment} object or one of its subclasses.
Rows correspond to genomic features and columns correspond to cells.}

\item{block}{Arguments passed to \code{\link[scrapper]{correctMnn}}.}

\item{more.mnn.args}{Named list of additional arguments to pass to \code{\link[scrapper]{correctMnn}}.}

\item{reddim.type}{String or integer specifying the \code{\link[SingleCellExperiment]{reducedDim}} entry on which to perform MNN correction.}

\item{output.name}{String containing the name of the \code{\link[SingleCellExperiment]{reducedDim}} entry in which to store the corrected embedding.}

\item{delayed.transpose}{Logical scalar indicating whether to delay the transposition when storing coordinates in the \code{\link[SingleCellExperiment]{reducedDims}}.}
}
\value{
\code{x} is returned with the corrected embedding stored as a \code{reducedDim} entry.
}
\description{
Correct batch effects from an existing embedding with mutual nearest neighbors (MNNs).
}
\examples{
sce <- getTestRnaData.se("pca")
# Treating the tissue of origin as the batch.
sce <- correctMnn.se(sce, sce$tissue)
reducedDimNames(sce)

}
\seealso{
\code{\link[scrapper]{correctMnn}} from the \pkg{scrapper} package.
}
\author{
Aaron Lun
}
