% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTestData.se.R
\name{getTestData.se}
\alias{getTestData.se}
\alias{getTestRnaData.se}
\alias{getTestAdtData.se}
\title{Get a test scRNA-seq dataset}
\usage{
getTestRnaData.se(at = "start")

getTestAdtData.se(at = "start")
}
\arguments{
\item{at}{String specifying the level of processing.}
}
\value{
A \link[SingleCellExperiment]{SingleCellExperiment} containing a dataset at the specified level of processing.

For \code{getTestRnaData}, this is a scRNA-seq dataset of the mouse brain,
where the main experiment contains RNA counts and the alternative experiments contain ERCC and repeat element counts.

For \code{getTestAdtData}, this is a CITE-seq dataset of human PBMCs,
where the main experiment contains RNA counts and the alternative experiment contains ADT counts.
}
\description{
Get a test scRNA-seq or CITE-seq dataset with varying levels of processing.
This uses caching to avoid recomputation.
}
\examples{
getTestRnaData.se()
getTestAdtData.se()

}
\seealso{
\code{\link[scRNAseq]{ZeiselBrainData}}, for the original data source for \code{getTestRnaData.se}.

\code{\link[scRNAseq]{KotliarovPBMCData}}, for the original data source for \code{getTestAdtData.se}.
}
\author{
Aaron Lun
}
