% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeRnaCounts.se.R
\name{normalizeRnaCounts.se}
\alias{normalizeRnaCounts.se}
\title{Normalize RNA counts}
\usage{
normalizeRnaCounts.se(
  x,
  size.factors = NULL,
  center = TRUE,
  block = NULL,
  mode = "lowest",
  log = TRUE,
  pseudo.count = 1,
  assay.type = "counts",
  output.name = "logcounts"
)
}
\arguments{
\item{x}{A \link[SummarizedExperiment]{SummarizedExperiment} object or one of its subclasses.
Rows correspond to genes and columns correspond to cells.}

\item{size.factors}{Numeric vector of length equal to the number of columns of \code{x},
containing the size factor for each cell in \code{x}.
If \code{NULL}, this defaults to the column sums of the count matrix in \code{x}.}

\item{center}{Logical scalar indicating whether to center the \code{size.factors},
see \code{?\link[scrapper]{centerSizeFactors}} for more details.}

\item{block, mode}{Arguments passed to \code{\link[scrapper]{centerSizeFactors}}.}

\item{log, pseudo.count}{Arguments passed to \code{\link[scrapper]{normalizeCounts}}.}

\item{assay.type}{Integer or string specifying the assay of \code{x} with the count matrix.}

\item{output.nme}{String containing the name of the assay to store the normalized matrix.}
}
\value{
\code{x} is returned with a new assay containing the (log-)normalized matrix.
Size factors are also stored in the \code{\link[SummarizedExperiment]{colData}}.
}
\description{
Compute (log-)normalized expression values after performing scaling normalization of an RNA count matrix.
}
\examples{
sce <- getTestRnaData.se("qc")
sce <- normalizeRnaCounts.se(sce, size.factors=sce$sum)
assayNames(sce)
summary(sizeFactors(sce))

}
\seealso{
\code{\link[scrapper]{centerSizeFactors}} and \code{\link[scrapper]{normalizeCounts}}, from the \pkg{scrapper} package.
}
\author{
Aaron Lun
}
