% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPca.se.R
\name{runPca.se}
\alias{runPca.se}
\title{Principal components analysis}
\usage{
runPca.se(
  x,
  features,
  number = 25,
  block = NULL,
  num.threads = 1,
  more.pca.args = list(),
  assay.type = "logcounts",
  prefix = NULL,
  output.name = "PCA"
)
}
\arguments{
\item{x}{A \link[SummarizedExperiment]{SummarizedExperiment} object or one of its subclasses.
Rows correspond to genomic features and columns correspond to cells.}

\item{features}{Integer, logical or character vector containing the features of interest to use in the PCA.
For RNA data, this is typically the \code{hvg} vector added by \code{\link{chooseRnaHvgs.se}}.
If \code{NULL}, all available features are used.}

\item{number, block, num.threads}{Arguments to pass to \code{\link[scrapper]{runPca}}.}

\item{more.pca.args}{Named list of additional arguments to pass to \code{\link[scrapper]{runPca}}.}

\item{assay.type}{Integer or string specifying the assay of \code{x} to be used for PCA.
This is typically the log-normalized expression matrix created by \code{\link{normalizeRnaCounts.se}}.}

\item{prefix}{String containing a prefix to add to the names in the output \code{link[SummarizedExperiment]{metadata}}.}

\item{output.name}{String containing the name of the components in the output \code{\link[SingleCellExperiment]{reducedDim}}.}
}
\value{
\code{x} is returned with the principal component scores in the \code{reducedDim}.
Additional information (rotation matrix, variance explained) is stored in the \code{\link[SummarizedExperiment]{metadata}}.
}
\description{
Compact and denoise the dataset by performing PCA on the (log-)normalized expression matrix.
}
\examples{
sce <- getTestRnaData.se("hvg")
sce <- runPca.se(sce, rowData(sce)$hvg)
dim(reducedDim(sce, "PCA"))
plot(metadata(sce)$variance.explained / metadata(sce)$total.variance)

}
\seealso{
\code{\link[scrapper]{runPca}} from the \pkg{scrapper} package.
}
\author{
Aaron Lun
}
