% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTsne.se.R
\name{runTsne.se}
\alias{runTsne.se}
\title{t-stochastic neighbor embedding}
\usage{
runTsne.se(
  x,
  perplexity = 30,
  num.threads = 1,
  more.tsne.args = list(),
  reddim.type = "PCA",
  output.name = "TSNE"
)
}
\arguments{
\item{x}{A \link[SingleCellExperiment]{SingleCellExperiment} object or one of its subclasses.
Rows correspond to genomic features and columns correspond to cells.}

\item{perplexity, num.threads}{Arguments to pass to \code{\link[scrapper]{runTsne}}.}

\item{more.tsne.args}{Named list of further arguments to pass to \code{\link[scrapper]{runTsne}}.}

\item{reddim.type}{Integer or string specifying the existing embedding in the \code{\link[SingleCellExperiment]{reducedDim}} of \code{x}.}

\item{output.name}{String containing the name of the output \code{\link[SingleCellExperiment]{reducedDim}}.}
}
\value{
\code{x} is returned with the t-SNE coordinates stored in the \code{reducedDim}.
}
\description{
Generate a t-SNE visualization from an existing embedding.
}
\examples{
sce <- getTestRnaData.se("pca")
# Using fewer iterations for a faster-running example.
sce <- runTsne.se(sce, more.tsne.args=list(max.iterations=50))
head(reducedDim(sce, "TSNE"))
 
}
\author{
Aaron Lun
}
