% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleByNeighbors.se.R
\name{scaleByNeighbors.se}
\alias{scaleByNeighbors.se}
\title{Scale and combine multiple embeddings}
\usage{
scaleByNeighbors.se(
  x,
  altexp.reddims,
  main.reddims = "PCA",
  num.neighbors = 20,
  num.threads = 1,
  more.scale.args = list(),
  output.name = "combined",
  meta.name = "combined",
  delayed.transpose = FALSE
)
}
\arguments{
\item{x}{A \link[SingleCellExperiment]{SingleCellExperiment} object or one of its subclasses.
Rows correspond to genomic features and columns correspond to cells.}

\item{altexp.reddims}{Named list of character or integer vectors.
Each entry is named after an alternative experiment.
Each vector contains the names/indices of the \code{\link[SingleCellExperiment]{reducedDim}} embeddings from that experiment to be combined.}

\item{main.reddims}{Character or integer vector specifying the names/indices of the \code{\link[SingleCellExperiment]{reducedDim}} entries from \code{x} to be combined.}

\item{num.neighbors, num.threads}{Arguments to pass to \code{\link[scrapper]{scaleByNeighbors}}.}

\item{more.scale.args}{Named list of additional arguments to pass to \code{\link[scrapper]{scaleByNeighbors}}.}

\item{output.name}{String containing the name of the \code{\link[SingleCellExperiment]{reducedDim}} entry in which to store the combined embeddings.}

\item{meta.name}{String containing the name of the \code{\link[S4Vectors]{metadata}} entry in which to store additional metrics.
If \code{NULL}, additional metrics are not stored.}

\item{delayed.transpose}{Logical scalar indicating whether to delay the transposition when storing coordinates in the \code{\link[SingleCellExperiment]{reducedDims}}.}
}
\value{
\code{x} is returned with the combined embeddings stored in its \code{rowData}.
The scaling factors for all embeddings are stored in the \code{metadata}.
}
\description{
Scale embeddings for different modalities to equalize their intra-population variance,
and then combine them into a single embedding for downstream analysis.
}
\examples{
sce <- getTestAdtData.se("pca")
sce <- scaleByNeighbors.se(sce, altexp.reddims=list(ADT="PCA"))
reducedDimNames(sce) 
metadata(sce)$combined

}
\seealso{
\code{\link{scaleByNeighbors}} from the \pkg{scrapper} package.
}
\author{
Aaron Lun
}
