% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickCrisprQc.se.R
\name{quickCrisprQc.se}
\alias{quickCrisprQc.se}
\alias{formatComputeCrisprQcMetricsResult}
\title{Quick quality control for CRISPR data}
\usage{
quickCrisprQc.se(
  x,
  num.threads = 1,
  num.mads = 3,
  block = NULL,
  assay.type = "counts",
  prefix = NULL,
  flatten = TRUE
)

formatComputeCrisprQcMetricsResult(compute.res, prefix = NULL, flatten = TRUE)
}
\arguments{
\item{x}{A \link[SummarizedExperiment]{SummarizedExperiment} object or one of its subclasses.
Rows correspond to CRISPR guides and columns correspond to cells.}

\item{num.threads}{Arguments passed to \code{\link[scrapper]{computeCrisprQcMetrics}}.}

\item{num.mads, block}{Arguments passed to \code{\link[scrapper]{suggestCrisprQcThresholds}}.}

\item{assay.type}{Integer or string specifying the assay of \code{x} containing the CRISPR count matrix.}

\item{prefix}{String containing a prefix to append to the name of each column corresponding to a QC metric in the \code{link[SummarizedExperiment]{colData}}.}

\item{flatten}{Logical scalar indicating whether to flatten the subset proportions into separate columns of the \code{link[SummarizedExperiment]{colData}}.
If \code{FALSE}, the subset proportions are stored in a nested \link[S4Vectors]{DataFrame}.}

\item{compute.res}{List returned by \code{\link[scrapper]{computeCrisprQcMetrics}}.}
}
\value{
For \code{quickCrisprQc.se}, \code{x} is returned with additional columns added to its \code{\link[SummarizedExperiment]{colData}}.
Each column contains per-cell values for one of the QC metrics, see \code{\link[scrapper]{computeCrisprQcMetrics}} for details.
The suggested thresholds are stored as a list in \code{\link[S4Vectors]{metadata}}.
The \code{colData} also contains a \code{keep} column, specifying which cells are to be retained.

For \code{formatComputeCrisprQcMetricsResult}, a \link[S4Vectors]{DataFrame} is returned with the per-cell QC metrics.
}
\description{
Quickly compute quality control (QC) metrics, thresholds and filters from CRISPR data in a \link[SummarizedExperiment]{SummarizedExperiment}.
}
\examples{
sce <- altExp(getTestCrisprData.se(), "CRISPR Guide Capture")
sce <- quickCrisprQc.se(sce)
colData(sce)[,c("sum", "detected", "max.value", "max.index")]
metadata(sce)$thresholds
summary(sce$keep)

}
\seealso{
\code{\link[scrapper]{computeCrisprQcMetrics}},
\code{\link[scrapper]{suggestCrisprQcThresholds}},
and \code{\link[scrapper]{filterCrisprQcMetrics}}
from the \pkg{scrapper} package.
}
\author{
Aaron Lun
}
