% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickRnaQc.se.R
\name{quickRnaQc.se}
\alias{quickRnaQc.se}
\alias{formatComputeRnaQcMetricsResult}
\title{Quick quality control for RNA data}
\usage{
quickRnaQc.se(
  x,
  subsets,
  num.threads = 1,
  num.mads = 3,
  block = NULL,
  altexp.proportions = NULL,
  assay.type = "counts",
  output.prefix = NULL,
  thresholds.name = "thresholds",
  flatten = TRUE
)

formatComputeRnaQcMetricsResult(compute.res, flatten = TRUE)
}
\arguments{
\item{x}{A \link[SummarizedExperiment]{SummarizedExperiment} object or one of its subclasses.
Rows correspond to genes and columns correspond to cells.}

\item{subsets, num.threads}{Arguments passed to \code{\link[scrapper]{computeRnaQcMetrics}}.}

\item{num.mads, block}{Arguments passed to \code{\link[scrapper]{suggestRnaQcThresholds}}.}

\item{altexp.proportions}{Unnamed integer or character vector containing the indices/names of alternative experiments for which to compute QC metrics, see Details.
The assay to use from each alternative experiment is determined by \code{assay.type}.

Alternatively, a named integer or character vector.
Each name specifies an alternative experiment while each value is the index/name of the assay to use from that experiment.

Only relevant if \code{x} is a \link[SingleCellExperiment]{SingleCellExperiment}.}

\item{assay.type}{Integer or string specifying the assay of \code{x} containing the RNA count matrix.}

\item{output.prefix}{String containing a prefix to add to the names of the \code{link[SummarizedExperiment]{colData}} columns containing the output statistics.}

\item{thresholds.name}{String containing the name of the \code{\link[S4Vectors]{metadata}} entry containing the filtering thresholds.
If \code{NULL}, thresholds are not stored in the metadata.}

\item{flatten}{Logical scalar indicating whether to flatten the subset proportions into separate columns of the \code{link[SummarizedExperiment]{colData}}.
If \code{FALSE}, the subset proportions are stored in a nested \link[S4Vectors]{DataFrame}.}

\item{compute.res}{List returned by \code{\link[scrapper]{computeRnaQcMetrics}}.}
}
\value{
For \code{quickRnaQc.se}, \code{x} is returned with additional columns added to its \code{\link[SummarizedExperiment]{colData}}.
Each column contains per-cell values for one of the QC metrics, see \code{\link[scrapper]{computeRnaQcMetrics}} for details.
The suggested thresholds are stored as a list in \code{\link[S4Vectors]{metadata}}.
The \code{colData} also contains a \code{keep} column, specifying which cells are to be retained.
If \code{altexp.proportions} is provided, QC metrics are added to the \code{colData} of the specified alternative experiments in the output object.

For \code{computeRnaQcMetricsWithAltExps}, a list is returned containing:
\itemize{
\item \code{main}, the result of calling \code{\link[scrapper]{computeRnaQcMetrics}} on the RNA count matrix in \code{x}.
The proportion of counts in each alternative experiment is added to the \code{subsets}.
\item \code{altexp}, a named list of length equal to \code{altexp.proportions}.
Each inner list is the result of calling \code{computeRnaQcMetrics} on the RNA count matrix of the corresponding alternative experiment of \code{x}.
}

For \code{formatComputeRnaQcMetricsResult}, a \link[S4Vectors]{DataFrame} is returned containing the per-cell QC metrics.
}
\description{
Quickly compute quality control (QC) metrics, thresholds and filters from RNA data in a \link[SummarizedExperiment]{SummarizedExperiment}.
}
\details{
\code{altexp.proportions} is typically used to refer to alternative experiments holding spike-in data.
For each alternative experiment, the proportion is defined \deqn{X/(X+Y)} where \deqn{X} is the alternative experiment's total and \deqn{Y} is the RNA total.
These proportions will be used for filtering in the same manner as the proportions computed from \code{subsets}.
}
\examples{
sce <- getTestRnaData.se()
sce <- quickRnaQc.se(sce, subsets=list(mito=grepl("^mt", rownames(sce))))
colData(sce)[,c("sum", "detected", "mito.proportion")]
metadata(sce)$thresholds
summary(sce$keep)

# Computing spike-in proportions, if available.
sce <- getTestRnaData.se()
sce <- quickRnaQc.se(
   sce,
   subsets=list(mito=grepl("^mt", rownames(sce))),
   altexp.proportions="ERCC"
)
colData(sce)[,c("sum", "detected", "mito.proportion", "ERCC.proportion")]
colData(altExp(sce, "ERCC"))[,c("sum", "detected")]

}
\seealso{
\code{\link[scrapper]{computeRnaQcMetrics}},
\code{\link[scrapper]{suggestRnaQcThresholds}},
and \code{\link[scrapper]{filterRnaQcMetrics}}
from the \pkg{scrapper} package.
}
\author{
Aaron Lun
}
